also filter atAddress manually in web.eth.filter
I successfully reverse engineered the possible assets via the correct abi
there is: EuroToken, BitcoinToken, RepToken and MelonToken. But still, with the
web3.eth.contract(abi) lookup, the fields of atTimestamp and ofPrice where empty.

But I realized, that the indexed inputs are in data.topics. And because ofAsset is
indexed, I could filter for the assets manually.